Split state_demo apart; rename to state_credits
