+++ SAUCE/defines/menunametable3.h
@@ -1,9 +1,3 @@
-
-CODE_BANK_PUSH("XCD_BANK_02")
-
-void mouse_and_cursor();
-void set_settings();
-
 const unsigned char instructions[448]={
 0x04,0xfe,0x04,0x3f,0xff,0x04,0x07,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28,
 0x29,0x22,0x23,0x22,0x2b,0x2c,0x2c,0xff,0x04,0x0f,0x30,0x31,0x32,0x33,0x34,0x35,
@@ -226,25 +220,6 @@ const unsigned char Credits2[629]={
 0x01,0x0d,0x00,0x01,0x00
 };
 
-
-#endif
-
-
-
-
-
-
-
-
-
-#include "charmaps/mainmenu_charmap.h"
-
-#ifdef FLAG_ENABLE_VER_NUM
-	#ifdef FLAG_BETA_BUILD
-		const unsigned char ver[] = "DEMO";
-	#else
-		const unsigned char ver[] = " VER";
-	#endif
 #endif
 
 #if !__VS_SYSTEM
@@ -252,417 +227,3 @@ const unsigned char palette_Credits[16]={ 0x11,0x0f,0x10,0x30,0x11,0x0f,0x2a,0x3
 #else
 const unsigned char palette_Credits[16]={ 0x06,0x0f,0x10,0x30,0x05,0x0f,0x2a,0x39,0x05,0x28,0x17,0x0f,0x05,0x0f,0x05,0x21 };
 #endif
-
-
-
-void state_instructions(){
-	mmc3_disable_irq();
-	
-	ppu_off();
-	pal_bright(0);
-    
-	pal_bg(palette_Credits);
-	
-
-
-	vram_adr(NAMETABLE_A);
-	vram_unrle(instructions);
-
-	// __asm__("LDA mmc3PRG1Bank \nPHA ");
-    // mmc3_set_prg_bank_1(0);
-    // vram_adr(NAMETABLE_A);
-	//    vram_unrle(dem_funnies);
-    // __asm__("PLA \n JSR %v ", mmc3_set_prg_bank_1);
-
-	oam_clear();
-	mmc3_set_8kb_chr(MENUBANK);
-	mmc3_set_2kb_chr_bank_0(0xFF);	
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);	
-	ppu_on_all();
-	music_update();
-	pal_fade_to(0,4);
-	tmp1 = 0;
-
-	/*	Incomplete code for reproducing what is being screamed into the Famicom microphone
-	(He forgor about the PCM bit)
-	OG by UserSniper, commit 7e47f425
-	POKE(0x4015, 0b00010000);
-	while (1) {
-		POKE(0x4011, fc_mic_poll()<<4);
-	}
-	*/
-
-	
-	do {
-		oam_clear();
-		mouse_and_cursor();
-		music_update();
-		ppu_wait_nmi();
-		newrand();
-		kandoframecnt++;
-	} while (!joypad1.press
-		#if !__VS_SYSTEM
-			&& !mouse.left_press
-		#endif
-		);
-	gameState = STATE_MENU;
-	return;
-}	
-
-void state_demo(){
-	mmc3_disable_irq();
-	
-    ppu_off();
-	pal_bright(0);
-    
-    pal_bg(palette_Credits);
-	
-
-
-	vram_adr(NAMETABLE_A);
-	vram_unrle(Credits);
-	vram_adr(NAMETABLE_B);
-	vram_unrle(Credits2);
-
-	
-
-	multi_vram_buffer_horz(ver, sizeof(ver)-1, NTADR_A(1,24));
-	one_vram_buffer(FLAG_MAJ_VER, NTADR_A(6,24));
-	#ifdef FLAG_BETA_BUILD
-		one_vram_buffer(0x18, NTADR_A(7,24)); // dot
-		one_vram_buffer(FLAG_MIN_VER, NTADR_A(8,24));
-//		one_vram_buffer(0x18, NTADR_A(9,24)); // dot
-//		one_vram_buffer(FLAG_MAJ_VER, NTADR_A(10,24));
-	#endif
-	
-	// __asm__("LDA mmc3PRG1Bank \nPHA ");
-    // mmc3_set_prg_bank_1(0);
-    // vram_adr(NAMETABLE_A);
-	//    vram_unrle(dem_funnies);
-    // __asm__("PLA \n JSR %v ", mmc3_set_prg_bank_1);
-
-	oam_clear();
-	mmc3_set_8kb_chr(MENUBANK);
-	mmc3_set_2kb_chr_bank_0(0xFF);	
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);	
-	ppu_on_all();
-	ppu_wait_nmi();
-
-	pal_fade_to(0,4);
-	tmp1 = 0;
-
-	/*	Incomplete code for reproducing what is being screamed into the Famicom microphone
-	(He forgor about the PCM bit)
-	OG by UserSniper, commit 7e47f425
-	POKE(0x4015, 0b00010000);
-	while (1) {
-		POKE(0x4011, fc_mic_poll()<<4);
-	}
-	*/
-	ppu_wait_nmi();
-	
-	do {
-		oam_clear();
-       	mouse_and_cursor();
-		newrand();
-	    kandoframecnt++;
-		
-		if (!forced_credits &&
-			(joypad1.press || ((mouse.connected)
-				? mouse.left_press || mouse.right_press
-				: joypad2.press))) {
-					gameState = STATE_MENU; return;
-		}
-		ppu_wait_nmi();
-		#if __VS_SYSTEM
-			crossPRGBankJump0(check_if_music_stopped);
-		#endif
-		music_update();
-		tmp1++;
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;	
-	} while (tmp1 != 0);
-
-	tmp1 = 0;
-	do {
-		oam_clear();
-		mouse_and_cursor();
-		newrand();
-	    kandoframecnt++;
-		#if __VS_SYSTEM
-			crossPRGBankJump0(check_if_music_stopped);
-		#endif
-		music_update();
-		if (!forced_credits &&
-			(joypad1.press || ((mouse.connected)
-				? mouse.left_press || mouse.right_press
-				: joypad2.press))) {
-				gameState = STATE_MENU; return;
-		}
-		
-		ppu_wait_nmi();
-		tmp1++;
-		set_scroll_x(tmp1<<2);
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;				
-	} while (tmp1 < 64);
-
-	tmp1 = 0;
-	set_scroll_x(256);
-	do {
-		oam_clear();
-		mouse_and_cursor();
-		newrand();
-		kandoframecnt++;
-		if (!forced_credits &&
-			(joypad1.press || ((mouse.connected)
-				? mouse.left_press || mouse.right_press
-				: joypad2.press))) {
-					gameState = STATE_MENU; return;
-		}
-		ppu_wait_nmi();
-		#if __VS_SYSTEM
-			crossPRGBankJump0(check_if_music_stopped);
-		#endif
-		music_update();
-		tmp1++;
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;				
-	} while (tmp1 != 0);
-	forced_credits = 0;
-	gameState = STATE_MENU;
-	return; 
-	
-}
-
-
-#include "charmaps/mainmenu_charmap.h"
-
-void settings() {
-	settingvalue = 0; 
-	pal_fade_to_withmusic(4,0);
-	mmc3_disable_irq();
-	ppu_off();
-	pal_bg(paletteSettings);
-	vram_adr(NAMETABLE_A);
-	vram_unrle(settingscreen);   	
-	mmc3_set_2kb_chr_bank_0(0xFF);
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);
-	ppu_on_all();
-	#if !__VS_SYSTEM
-	one_vram_buffer('c', NTADR_A(4, 5));	// settingvalue is set to 0 beforehand
-	#else
-	one_vram_buffer('c', NTADR_A(4, 7));	// settingvalue is set to 0 beforehand
-	#endif
-	pal_fade_to_withmusic(0,4);
-	while (1) {
-		ppu_wait_nmi();
-		music_update();
-		oam_clear();
-		mouse_and_cursor();
-		 // read the first controller
-		
-	#if !__VS_SYSTEM
-		if (twoplayer) one_vram_buffer('g', NTADR_A(26, 5));
-		else one_vram_buffer('f', NTADR_A(26, 5));
-	#else
-		if (twoplayer) one_vram_buffer('g', NTADR_A(26, 7));
-		else one_vram_buffer('f', NTADR_A(26, 7));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (options & oneptwoplayer) one_vram_buffer('g', NTADR_A(26, 7));
-		else one_vram_buffer('f', NTADR_A(26, 7));
-	#else
-		if (options & oneptwoplayer) one_vram_buffer('g', NTADR_A(26, 9));
-		else one_vram_buffer('f', NTADR_A(26, 9));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (options & sfxoff) one_vram_buffer('f', NTADR_A(26, 9));
-		else one_vram_buffer('g', NTADR_A(26, 9));
-	#else
-		if (options & sfxoff) one_vram_buffer('f', NTADR_A(26, 11));
-		else one_vram_buffer('g', NTADR_A(26, 11));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (options & musicoff) one_vram_buffer('f', NTADR_A(26, 11));
-		else one_vram_buffer('g', NTADR_A(26, 11));
-	#else
-		if (options & musicoff) one_vram_buffer('f', NTADR_A(26, 13));
-		else one_vram_buffer('g', NTADR_A(26, 13));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (options & jumpsound) one_vram_buffer('g', NTADR_A(26, 13));
-		else one_vram_buffer('f', NTADR_A(26, 13));
-	#else
-		if (options & jumpsound) one_vram_buffer('g', NTADR_A(26, 15));
-		else one_vram_buffer('f', NTADR_A(26, 15));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (viseffects) one_vram_buffer('g', NTADR_A(26, 15));
-		else one_vram_buffer('f', NTADR_A(26, 15));
-	#else
-		if (viseffects) one_vram_buffer('g', NTADR_A(26, 17));
-		else one_vram_buffer('f', NTADR_A(26, 17));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (trails == 1) one_vram_buffer('g', NTADR_A(26, 17));
-		else if (trails == 2) one_vram_buffer('*', NTADR_A(26, 17));
-		else one_vram_buffer('f', NTADR_A(26, 17));
-	#else
-		if (trails == 1) one_vram_buffer('g', NTADR_A(26, 19));
-		else if (trails == 2) one_vram_buffer('*', NTADR_A(26, 19));
-		else one_vram_buffer('f', NTADR_A(26, 19));
-	#endif
-
-	#if !__VS_SYSTEM
-		if (auto_practicepoints) one_vram_buffer('g', NTADR_A(26, 19));
-		else one_vram_buffer('f', NTADR_A(26, 19));
-	#endif
-
-		tmp1 = settingvalue;
-
-		if (mouse.left_press) {
-			if (mouse.x >= 0x2D && mouse.x <= 0xDD) {
-			#if !__VS_SYSTEM
-				if (mouse.y >= 0x24 && mouse.y <= 0x2C) {
-					settingvalue = 0; set_settings();
-				}
-				else if (mouse.y >= 0x34 && mouse.y <= 0x4C) {
-					settingvalue = 1; set_settings();
-				}
-				else if (mouse.y >= 0x44 && mouse.y <= 0x4C) {
-					settingvalue = 2; set_settings();
-				}
-				else if (mouse.y >= 0x54 && mouse.y <= 0x5C) {
-					settingvalue = 3; set_settings();
-				}
-				else if (mouse.y >= 0x64 && mouse.y <= 0x6C) {
-					settingvalue = 4; set_settings();
-				}
-				else if (mouse.y >= 0x74 && mouse.y <= 0x7C) {
-					settingvalue = 5; set_settings();
-				}
-				else if (mouse.y >= 0x84 && mouse.y <= 0x8C) {
-					settingvalue = 6; set_settings();
-				}
-				else if (mouse.y >= 0x94 && mouse.y <= 0x9C) {
-					settingvalue = 7; set_settings();
-				}
-			#else
-				if (mouse.y >= 0x34 && mouse.y <= 0x3C) {
-					settingvalue = 0; set_settings();
-				}
-				else if (mouse.y >= 0x44 && mouse.y <= 0x4C) {
-					settingvalue = 1; set_settings();
-				}
-				else if (mouse.y >= 0x54 && mouse.y <= 0x5C) {
-					settingvalue = 2; set_settings();
-				}
-				else if (mouse.y >= 0x64 && mouse.y <= 0x6C) {
-					settingvalue = 3; set_settings();
-				}
-				else if (mouse.y >= 0x74 && mouse.y <= 0x7C) {
-					settingvalue = 4; set_settings();
-				}
-				else if (mouse.y >= 0x84 && mouse.y <= 0x8C) {
-					settingvalue = 5; set_settings();
-				}
-				else if (mouse.y >= 0x94 && mouse.y <= 0x9C) {
-					settingvalue = 6; set_settings();
-				}
-			#endif
-
-			}
-			if ((mouse.x >= 0x1D && mouse.x <= 0xDD) && (mouse.y >= 0xBC && mouse.y <= 0xC4)) {		
-				return;
-			}
-
-		}	
-
-		if (joypad1.press & (PAD_RIGHT | PAD_DOWN)) {
-		#if !__VS_SYSTEM
-			if (settingvalue == 8) { settingvalue = 0;  }
-		#else
-			if (settingvalue == 7) { settingvalue = 0;  }
-		#endif
-			else { settingvalue++;   }
-		}
-
-		if (joypad1.press & (PAD_LEFT | PAD_UP)) {
-		#if !__VS_SYSTEM
-			if (settingvalue == 0) { settingvalue = 8;  }
-		#else
-			if (settingvalue == 0) { settingvalue = 7;  }
-		#endif
-			else { settingvalue--;   }
-		}
-
-		if (tmp1 != settingvalue) {
-			// NTADR_A = (NAMETABLE_A|(((y)<<5)|(x)))
-			// (tmp1 * 2) << 5 = tmp1<<6 = (tmp1<<8)>>2
-		#if !__VS_SYSTEM
-			one_vram_buffer(' ', NTADR_A(4, 5)+((tmp1<<8)>>2));
-			one_vram_buffer('c', NTADR_A(4, 5)+((settingvalue<<8)>>2));
-		#else
-			one_vram_buffer(' ', NTADR_A(4, 7)+((tmp1<<8)>>2));
-			one_vram_buffer('c', NTADR_A(4, 7)+((settingvalue<<8)>>2));
-		#endif
-		}
-		
-		if (joypad1.press & (PAD_A | PAD_START)) {
-			set_settings();
-		}
-		if (options & platformer) {
-			twoplayer = 0;
-			one_vram_buffer('X',NTADR_A(26,7));
-		}
-		if (twoplayer) options &= ~platformer;		
-
-		if (joypad1.press_b) {
-			return;
-		}
-		kandoframecnt++;
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;	
-				
-	}
-}
-
-const uint8_t xbgmlookuptable[];
-
-void set_settings() {
-	switch (settingvalue) {
-		case 0: // oneptwoplayer
-			twoplayer ^= 1; break;
-		case 1: // oneptwoplayer
-			options ^= oneptwoplayer; break;
-		case 2: // sfxoff
-			options ^= sfxoff; break;
-		case 3: // musicoff
-			#if __VS_SYSTEM
-			options ^= musicoff; if (options & musicoff) { famistudio_music_stop(); music_update(); } else { music_play(idx8_load(xbgmlookuptable, newrand() & 31)); } break;
-			#else
-			options ^= musicoff; if (options & musicoff) { famistudio_music_stop(); music_update(); } else { music_play(xbgmlookuptable[menu_music]); } break;
-			#endif
-		case 4: // jumpsound
-			options ^= jumpsound; break;
-		case 5:
-			viseffects ^= 1; if (!viseffects) trails = 0; break;
-		case 6:
-			trails = trails == 2 ? 0 : trails + 1; break;					
-		case 7:
-	#if !__VS_SYSTEM
-			auto_practicepoints ^= 1; break;
-		case 8:
-	#endif
-			if (joypad1.a && joypad1.press_start) {
-				setdefaultoptions();
-				__asm__("JMP ($FFFC)");	// restart the game lmao	
-			}
-			break;
-	}
-}	
-
-CODE_BANK_POP()
